refactor: entity 클래스들 리팩토링
id필드의 타입을 int에서 Long으로 변경함
@Table, @Column을 추가하여 테이블의 컬럼명을 바로 알 수 있도록 가독성을
높임

+++ FinalToyProject/src/main/java/com/fastcampus/biz/persistence/BlogRepository.java
@@ -9,7 +9,7 @@ import com.fastcampus.biz.domain.Blog;
 
 
 @Repository
-public interface BlogRepository extends JpaRepository<Blog, Integer> {
+public interface BlogRepository extends JpaRepository<Blog, Long> {
 
 	List<Blog> findByTitleContains(String title);
 	List<Blog> findByTagContains(String tag);
