+++ FinalToyProject/src/main/java/com/fastcampus/biz/service/PostService.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Service;
 
 import com.fastcampus.biz.domain.Post;
 import com.fastcampus.biz.persistence.PostRepository;
-import com.fastcampus.web.domain.RequestPost;
+import com.fastcampus.web.dto.RequestPost;
 
 import lombok.RequiredArgsConstructor;
 
@@ -18,15 +18,15 @@ public class PostService {
 
 	private final PostRepository postRepository;
 	
-	public Post getPost(int postId) {
+	public Post getPost(Long postId) {
 		return postRepository.findById(postId).get();
 	}
 	
-	public List<Post> getPosts(int blogId) {
+	public List<Post> getPosts(Long blogId) {
 		return postRepository.findByBlogId(blogId);
 	}
 	
-	public List<Post> getPostsByCategoryId(int categoryId) {
+	public List<Post> getPostsByCategoryId(Long categoryId) {
 		return postRepository.findByCategoryId(categoryId);
 	}
 	
@@ -41,7 +41,7 @@ public class PostService {
 	}
 	
 	@Transactional
-	public void updatePost(int postId, RequestPost requestPost) {
+	public void updatePost(Long postId, RequestPost requestPost) {
 		Post post = postRepository.findById(postId).get();
 		post.setCategoryId(requestPost.getCategoryId());
 		post.setContent(requestPost.getContent());
@@ -49,11 +49,11 @@ public class PostService {
 	}
 	
 	@Transactional
-	public void deletePostByPostId(int postId) {
+	public void deletePostByPostId(Long postId) {
 		postRepository.deleteById(postId);
 	}
 	
-	public void deletePostByBlogId(int blogId) {
+	public void deletePostByBlogId(Long blogId) {
 		postRepository.deleteByBlogId(blogId);
 	}
 }
